ttools: modify use of rounding in BinSizer interface
The rounding requirements are now passed to a BinSizer as part
of the getWidth call rather than baked in at construction time.
This means that a plotter can use the same BinSizer config key
but be able to operate on plot surfaces with different rounding
requirements.

Different rounding is now performed according to whether the
axis for which bins are required is a time axis (in which case
rounding should be to units like hours/days/years) or a normal
numeric axis (in which case just use a round decimal number).
Nicer user experience for time-based histograms etc.