ttools: rewrite ChooserColorSpecifier
The basic behaviour is not changed from the previous version, but this
time it's based on a JPopupMenu rather than posting a JDialog which
is dismissed when the user clicks outside of the window.

This could be seen as abuse of JPopupMenu (the whole JColorChooser
becomes a menu item) it seems to be more in line with other Swing
components like JComboBox, both in terms of implementation
(no Toolkit-wide AWTEventListeners required) and in terms of
user experience.