table: add score scaling information to MatchEngine
Add a new method getScoreScale to the MatchEngine interface.

This is a non-backward-compatible change to the STIL API.
However, I'm not aware of implementations of this interface
outside of the starjava codebase, so it's probably not too harmful.
Implementation is in any case not arduous; it is permissible just
to return NaN from the method if no better value is available.

The point of this is to be able to fix the current bad behaviour of
the CombinedMatchEngine's getScore method.  Up till now, it has
returned a score based on simply adding together the scores
of its constituent matchers (e.g. arcsec + magnitudes).
While this does not contravene the MatchEngine contract,
it means that the numeric value of the score is not useful,
in particular it doesn't make any sense to look for a "closest"
match by choosing the lowest score.  So the "Best match" behaviour
of topcat and stilts up till now has been pretty questionable
for combined match engines.

Now we have a scale value for the scores, we can use dimensionless
values for the score of each of a CombinedMatchEngine's constituent
matchers, and combine them in quadrature.  Thus, CombinedMatchEngine
scores are now comparable.