ttools: fix broken withScroll plot padding switch
PlaneAxisAnnotation.getPadding has a withScroll argument which
should affect the amount of padding applied to the plot
(if the plot will not be scrolled, less padding may be needed).
This has been broken for a long while; the parameter was ignored
and the extra padding was applied anyway.

Fix it.