table: per-row error match engines now give scaled scores
The matchScore (and corresponding Info) for ErrorCartesianMatchEngine
and ErrorSkyMatchEngine now yield scaled values, that is
in the range 0..1, rather than in physical coordinates.
This makes more sense, since these scores can be compared
against each other.  A Best match makes more sense if the
results compared are scaled against the per-row errors rather
than absolute values.

This is a change of (undocumented) behaviour, but is only likely to
cause people trouble if they are looking at the output score values.