table: add spherical switch for CombinedMatchEngine
Francois-Xavier Pineau is of the opinion that the combined matcher
should only signal a match if the scaled separation of points in
N-dimensional space is within the unit sphere, rather than just
if all the consituent matchers indicate a match.
The difference corresponds to points in the "corners" of the
N-dimensional match volume.

Add a switch member variable inSphere that can change this behaviour.
At this commit, the behaviour is not changed.