ttools: improve plane axis annotation padding calculations
The PlaneAxisAnnotation.getPadding method, which calculates how
much space is required for drawing axis labels, is now done
using the same code as that which actually draws the axis labels,
which means it should get it right.  Previously the code was
complicated because the (anyway, complicated) logic was in two
different places, and it did indeed seem to get it wrong by a few
pixels in some cases (bottom edge of X axis labels truncated
at large font sizes).  Hopefully this should behave better,
and even if not it should be easier to fix correctly.

This commit actually obviates the need for (and erases the code from)
the previous one, but the previous commit was a bug fix, so
leave it in place for the record.