ttools: add Combiner config key to HistogramPlotter
The HistogramPlotter how has a Combiner key, allowing you to
control how values are aggregated within a histogram bin.
This only makes much sense for weighted bins.
The same capability is already present in the various 2-d
histogram/density map variants.

This was suggested by Henrik Rydberg.  I think it makes sense,
but there are a couple of concerns specific to using it with
1-d histograms.  One is the way it interacts with normalisation
and cumulative modes is hard to make sense of (for me anyway) -
probably for combiners other than the default (SUM) they don't
make much sense.  The other is that for consistency, probably
the same option should be introduced to the KDE-like plotters
(Pixel1dPlotter and concrete subclasses), this hasn't been done yet.