ttools: add example plot targets to build file
The ttools build.xml file now contains a new public target,
plot2-examples.  This uses the Plot2Example class to generate
example plots programmatically and insert them into the
source tree.  This target is not invoked as part of the normal
build process, it should be invoked by hand from time to time
when the plotting code is changed (especially before a release).

Note as well as regenerating the images, this also serves as
a nice regression test, since figures whose pixels have changed
since the last time this target was invoked will show up as
modified files in the source code repository.