topcat: improve PNG generation
There is a bug somewhere (probably FOP) that means generating PDFs
using PNG files with transparent black backgrounds (0x00000000)
can end up with (opaque) black bacgrounds in the PDF.
This shows up as ugly/illegible icons in the generated sun253.pdf file.

Supplying PNGs with transparent white backgrounds (0x00ffffff)
works round the problem, but it seems to be extremely difficult
to get a BufferedImage to have a background of 0x00ffffff
(e.g. setting the Composite to SRC and fillRect doesn't do it;
I tried a bunch of other things that failed too).
I previously had some steps in IconUtils that were supposed to
achive this, but I don't think were doing it at all.

So hide the relevant code in IconUtils, and write some custom code
instead in topcat.ResourceIcon, which needs to turn Icons
(actually ImageIcons) into BufferedImages.
This still isn't very satisfactory: it only works when the
original icon is an image with a transparent white background.