ttools: fix plot title placement bug
The title placement in a plot (if present) was not allocating
space for padding between the plot bounds and the title.
That meant that the top part of the title could get chopped off,
especially in the case of LaTeX captioning.

The padding may still not be quite right (inconsistent between
plain and LaTeX captioning; I think they take account of
font descenders differently), but at least the title placement
is behaving better now.