ttools: SubrangeConfigKey avoids direct use of JideRangeSlider
SubrangeConfigKey now uses the new RangeSliderUtil class to
create and work with range-capable sliders.  That means it will
work (on a best efforts basis) in environments where the JIDE-OSS
RangeSlider implementation fails, e.g. (at time of writing)
with certain look-and-feels such as Nimbus, and on OSX Java 10
where the WindowsLookAndFeel classes are absent.
Previously, either of those cases would cause an exception on
the EDT, meaning the TOPCAT plot windows just didn't work.