ttools: SubrangeConfigKey now OK with non-standard L&Fs
SubrangeConfigKey now uses the new RangeSliderUtil class to
create and work with range-capable sliders.  That means it will
work (on a best efforts basis) with non-JIDE-OSS-supported
look-and-feels such as Nimbus - it just falls back to a local
(somewhat inferior) RangeSlider implementation if necessary.
Previously, attempting to use this config key with Nimbus
installed led to a NullPointerException on the Event Dispatch Thread.