topcat: fix bug when saving plot-report-generated tables
There was a bug (reported by Ronald Drimmel) when saving tables
generated from a plot's Report, for instance HEALPix pixel value
tables generated from the SkyDensity plot.  The second time you
saved such a table, it saved the data from the first time you did it.

This was because SaveTableQueryWindows are cached for use by the
AuxWindow.createSaveTableAction method.  That's not such a bad
idea, since preserving and re-using the state from earlier save
dialogues is likely to be useful for the user.  But it was
saving and reusing the table data source as well.

Fix it so that a new TableSource is installed every time the
save window is used.  It's still not bulletproof, since you
might get strange behaviour if you are trying to save two tables
at the same time while one of the save dialogues is iconised or
something ... but future problems should be fairly unlikely.