fits: move check for BINTABLE output column count
When writing a FITS binary table, the check for whether the legal
column limit has been exceeded is now done when the FitsTableSerializer
is constructed (required in the contract of the abstract
AbstractFitsTableWriter.createSerializer method, in practice in
concrete FitsTableSerializer implementation constructors)
rather than in the AbstractFitsTableWriter.writeTableHDU implementation.

This means the check has to be done in more places, but it gives
different FITS TableWriter implementations the chance to implement
it differently.