vo: rewrite Aux[Tap]ServiceFinder
In consultation with Markus, rewrite AuxServiceFinder,
which is a TapServiceFinder implementation that works as intended
for the scheme described in the IVOA Note "Discovering Data
Collections Within Services", v1.1.  The ADQL query is quite a lot
different than the one used before, which I think corresponded
to v1.0 of that note.

This is still considered experimental.  It's only likely to be used
by registry enthusiasts for now, and misses lots of services and
tables given the current state of the registry, which for most
TAP/tableset resources is not properly set up for the proposed
convention.

tmp: aux replacement