ttools: implement correlation-based plane error ellipses
Add a new MultiPointForm to paint error ellipses based on axis-aligned
coordinate errors as well as correlation values.  The visual effect
is the same as the position-angle-based error ellipse drawing that's
already in place, but in this case the correlation values are used
to construct a covariance matrix, which in turn defines the ellipse
orientation.

This is convenient for Gaia data, which comes with pair-correlations
for various values (ra, dec, pmra, pmdec, parallax).

Not available from the UI at this commit.