ttools: make sky vector/ellipse plotters viewsys invariant
Fix the SkyVectorCoordSet and SkyEllipseCoordSet, hence sky
vector and ellipse plotters, so that they interpret the
delta-longitude and delta-latitude extra coordinates to be
in the coordinates of the input data rather than the coordinates
of the view sky system.  Previously e.g. a vector with
deltaLatitude=0 was always parallel with the longitude lines
of the view sky system, and so it could change direction on
the physical sky depending on whether you were looking at the
plot in galactic or ecliptic coordinates etc.  That's wasn't
unambiguously buggy, since it wasn't documented otherwise,
but it was probably not what people were expecting, and it's
not very useful.

Now the vectors and ellipses retain their positions on the
physical sky, regardless of the current view coordinate system.