ttools: hack Plot2Example to cope with multiple filter cmds
Add some code in Plot2Example to allow the example stilts invocations
to work if there are multiple filter parameters (cmd=... cmd=...)
strung together.  Otherwise it doesn't let you assign
multiple parameters with the same name.

This is a bit of a hack, since it does it by checking the parameter
name.  However, this code is only executed at build time, so it
doesn't need to be as robust as user-facing code.