ttools: replace plot2.data.CachedSequence with CachedReader
CachedColumn now provides a random-access rather than sequential-access
reader in order to access its data.  This is just an implementation
change local to the CachedDataStoreFactory implementation of
DataStoreFactory.  It is not expected to have much effect in terms
of performance, and shouldn't have any change to behaviour.
But it opens up the way to provide multiple iterators over different
parts of the cached data rather than a single stream.