ttools: split out Tuple interface from TupleSequence
Now a separate interface Tuple contains the accessor methods from
TupleSequence, and TupleSequence extends Tuple, adding only the next()
method.

This makes no change to behaviour, it just clarifies in some
other interfaces in the API that a supplied object is only
for interrogating but not for iterating over (since iteration is
done by some external component).