topcat: fixed export bug in Stats window
The Sum column in the StatisticsWindow's TableModel was incorrect:
it declared a contentClass of Double, but for boolean data table
columns it was yielding a Long.  That wasn't a problem for the
JTable display, but it led to a class incompatibility when
attempting to export the stats (save to disk or import as new table),
manifesting as a stacktrace preventing the export action.

Thanks to Henrik Rydberg for reporting this.