gbin: refactor GBIN table reading classes
Remove the class GbinTableReader and replace it with an abstract
class GbinStarTable which has the same functionality, including
some of the code that was previously in GbinTableBuilder.

This localises the logic required for turning GaiaRoot objects
into a StarTable in one place, GbinStarTable.
The GbinTableBuilder class now mostly just contains the I/O
required to harness that.

This is a more logical arrangement, and makes it easy to
add GbinStarTable implementations not based on a GBIN file
as such.  It wasn't done that way in the first place because
the reading is a bit fiddly, since you have to read the
first item from the stream to find out what class it has
before you can set up the table column metadata.
But it's done now.