ttools: rename TAPRegExt-v1.0.xsd resource file for clarity
The file TAPRegExt-v1.0.xsd is renamed to TAPRegExt-v1.0-pre-Erratum1
as a documentation of the fact that this is (following the acceptance
of TAPRegExt-v1.0-Erratum1) no longer the accepted correct content
of ththe TAPRegExt v1.0 schema.

This is purely a change in historical documentation; this resource
is in fact not used used by IvoaSchemaResolver (which uses instead
TAPRegExt-v1.0-Erratum1.xsd) or anywhere else in starjava code.