ttools: undeprecate VOTable 1.3 COOSYS (VOTable-1.3 Erratum 1)
The votlint version-sensitive VOTable validation code now does
not issue a deprecation warning if it finds a COOSYS element
in a VOTable 1.3 document.
If if finds a COOSYS in a VOTable 1.2 document, a warning is
still issued, but it points out that the deprecation is
revoked at 1.3.

This is in consequence of Erratum #1 to the VOTable 1.3 standard.