ttools: respect supplied sequence for metadata filter
The "meta" filter that is used for flexible display of per-column
metadata was using a hardcoded list of metadata items in order
to decide which order to write the columns in, if no sequence
was supplied from the command parameters.

However, it was ignoring the sequence provided (for just this
sort of purpose) by StarTable.getColumnAuxDataInfos() method.
Now it uses that sequence as well.