votable: VOTableFitsTableWriter now uses VOTableWriter
Replace code in the VOTableFitsTableWriter class with calls to
corresponding functionality in VOTableWriter.  This reduces
duplicated code and does more or less the same thing.
However, there are differences in the output serializations.

The default FITS-plus output now uses the default VOTable version
for metadata encoding, instead of VOTable 1.1 which it was using
before.  That's probably a good thing, but it's just possible that
it might cause problems elsewhere.