ttools: add multicolour interpolating shader
Add an overloaded Shaders.createInterpolatingShader() method that
can give you a shader that interpolates between any number of
colours (in pairs), not just two.

Behaviour of the 2-colour one is slightly changed: it now sets the
alpha component not just the RGB ones.  I think this is correcting
an oversight, but it's possible it could cause trouble.