ttools: use explicit auxmap aux/weighted shading examples
In the example plots for shading=aux and shading=weighted
set the auxmap parameter explicitly (in this case to plasma)
rather than using the default.  This is worth doing in
the stilts example command lines, to show how you can adjust
the colour map, since otherwise it's not very obvious from
the documentation (since aux* parameters are a property of
the plot command not the layer).