ttools: remove projection member from SkyAspect
SkyAspect does not need to contain the Projection at all;
that cannot change as a result of navigation actions,
and is properly managed by the SkySurfaceFactory's Profile
and not its Aspect.  So it is now removed, and it's clearer
what information SkyAspect aggregates.

Projection should never have been present in SkyAspect, so
this is just correcting a long-standing confusion. I hope.