jniast: Update astSlaAdd, astSpecAdd & astTimeAdd
Add the new "nargs" argument.  This is currently derived automatically
from the length of the argument array given.  The alternative would
be (as in the C-interface) to have the user provide the number of
arguments manually as an extra argument.

(cherry picked from commit 345e9d6364dd611475ba151617d895f0258a971a)