ttools: modify taplint examples @vocab requirements
Following DALI 1.1 and discussions with Markus and Pat, change the
taplint reporting on the @vocab attribute value in TAP examples
documents (EXA stage).

This is now complicated, because the requirements are a mess -
essentially, there is no way to write an examples document which
is standards-compliant and will work with the existing deployed
base of topcats out there (not entirely my fault).  Try to write
taplint reports comprehensive enough to explain to TAP service
providers what the situation is.