tamfits: set FitsFactory.useHierarch to true
Change the static setting that determines whether the nom.tam.fits
libraries recognise the ESO HIERARCH convention for hierarchical FITS
keywords (https://fits.gsfc.nasa.gov/registry/hierarch_keyword.html)
from false (which is how it is configured by default) to true.

This setting is required to make the hierarchical variant of the
WideFits convention work (see uk.ac.starlink.fits.WideFits).
There's no way in the nom.tam.fits API to configure HIERARCH-use
on a per-call basis, so we have to modify the static setting.

This will affect other nom.tam.fits-dependent code using this library.
An alternative would be to set this only when the hierarch convention
is potentially or actually in use, but that gets messier,
since it would still affect other code running in the same JVM,
but in less predictable ways.

So following this change, STIL FITS handling will pick up HIERARCH
per-table headers which it previously ignored.