fits: introduce framework for WideFits convention
Introduce the abstract machinery for using in BINTABLE-based
FITS tables  that allows storage of columns beyond the
BINTABLE-imposed maximum of 999 columns.

The 999 column limit comes from the fact that FITS keywords are
limited to 8 characters, and the per-column keywords are in general
of the form XXXXXnnn, where nnn gives the (1-based) column index.

This convention, implemented in the uk.ac.starlink.fits.WideFits class,
and described in src/docs/wide-fits.txt,
lays out the data exactly as it would for normal columns in the
data part of the HDU, but all columns beyond an extension column
index (usually 999) are included within a single container column
(999).  Header keywords are added somehow or other to define
the metadata for these extension columns.  At this commit,
no concrete proposal (AbstractWideFits concrete subclass)
is provided to do this.

The convention is outside of, but compatible with, the FITS
standard.  Wide tables making use of this convention will be
readable with standard FITS readers, but the standard FITS
readers will not see the columns beyond 999 as separate
columns.