fits: note relic status of base-26 WideFits
Following discussion on FITSBITS (thread
https://listmgr.nrao.edu/pipermail/fitsBIts/2017-July/002967.html,
and especially followups to
https://listmgr.nrao.edu/pipermail/fitsbits/2017-July/003007.html)
the HIERARCH variant of WideFits appears to be preferred
over the base-26 variant.

This commit notes that the base-26 variant is now chiefly of
historical interest.  The code is for now left harmlessly in place,
but if its maintenance becomes burdensome it could just be removed.