ttools: fix ErrorRenderer clip bug
ErrorRenderer used the graphics clip in a few places to work out
what (error-bound) offsets are a long way away from the central
position, in order not to try to plot kilometer-long lines and
stuff the AWT graphics system.  But this was based on a misconception:
the clip is not necessarily the size of the plot surface,
since if part of the plotting surface is obscured than AWT will
just present the newly-visible part as the clip.
In most cases for TOPCAT/STILTS this doesn't matter, since the
whole plot is rendered at once, but it can do, and one place it
does is when drawing the error renderer in a JComboBox, which
can get scrolled slowly onto the screen and have a clip that is
just a few pixels wide/high.

Fix this by using Graphics2D.getDeviceConfiguration().getBounds()
rather than Graphics2D.getClip().getBounds().

Also lets me remove a hack around the previous (bad) behaviour
to fix rendering in JComboBoxes, which anyway didn't work very
well since it ended up rendering on components overlaying the
combo box (like scrollbars).