ttools: add AngleUnit option in sky extent plotters
SkyVector, SkyEllipse and SkyCorr now have a Unit config option.
These let you specify the units in which the lon/lat displacements
are given.  The units can be degrees, arc minutes, arcsec etc,
including milliarcsec, which is used for gaia sky positional
errors.  If the given AngleUnit is null ("scaled") then autoscaling
is performed - this corresponds to the old autoscale option,
which is now withdrawn.  This therefore represents a backward
incompatibility, but the new config options should be easier to
use in most cases than the old ones, and are more flexible.