ttools: turn off vector/ellipse autoscaling by default
Change the default autoscaling configuration to FALSE
for all XY(Z) and Sky vector and ellipse plotters.

Previously it was true by default.  Setting autoscaling on by
default did give you a nice looking plot in most cases, but
if you weren't paying attention you cound end up thinking that
you had a quantitative rather than a qualitative plot, which
is a nasty mistake to make.  This way round, the plot you get
will probably be ugly to start with, but at least that way you
know you have to do something about it and have to think about
what that is.  I don't like to make these backwardly incompatible
changes between versions, but in this case I think I made the
wrong decision to start with.  Introducing the new XYCorr/SkyCorr
plotters, which behave differently, and adding the AngleUnit
scaling config for SkyEllipse and SkyVector, seems like a
reasonable time at which to make the change.