topcat: keep row highlighted when dragging in col window
When dragging a JTable row in the ColumnInfoWindow to change its
position in the TopcatModel column list, keep it selected
(highlighted).  That was supposed to happen anyway, but some
interaction of mouse listeners or something made it flickering
and mostly invisible.  Hack a fix by scheduling the row
selection later on the Event Dispatch Thread, which seems to work.