topcat: use null ControlWindow in CodecTest
Modify the CodecTest unit test so that it uses a null value
for the ControlWindow argument of the TopcatModel constructor.
TopcatModel really shouldn't require this (GUI-based) object,
ideally various things should be rewritten to avoid requiring
ControlWindow and thus tying them to the GUI.
But in the mean time, using a null value in this test doesn't
change the way anything behaves, and it means the test can
run headless.