table: update PixTools to 2017-09-06 version (BUGFIX)
This fixes the PixTools vect2pix_ring bug tested for in
the tests introduced at the previous commit.

Nikolay Kuropatkin (PixTools author) kindly fixed the bug quickly;
the replacement version introduced at this commit is from the git
repository at https://github.com/kuropat/eag-HEALPix,
commit 447a7be073876dba32d1c6be811eebc48255c844
(built from source by me).

It looks like this bug could in some cases affect the result of the
ring-scheme query_disc invocations on which the sky crossmatching
(PixtoolsHealpixSkyPixellator) is based - you can imagine it
failing to detect an association between two positions falling
either side of longitude=0 in one of the affected HEALPix tiles.
However, I've tried to write tests in which such failures would
manifest, and have not managed to come up with one.
So I *think* the practical impact of this bug for sky matches
would have been zero.  But I'm not 100% certain.

Even if there were some possible failures, given that the HEALPix
level is usually quite large for crossmatches (corresponding to
a small match radius), the scale of the problem (proportion of
the sky in which it could have an effect) would be pretty small.