topcat: reject SAMP table with duplicate table-id
If a table.load.* SAMP message is received with a table-id
message parameter duplicating that of a table loaded from an
earlier table.load.* message, an error response is now generated
and the second table is not loaded.  Previously it just loaded
it anyway.

This anomaly was pointed out by Tom Robitaille.