ttools: fix (unused) JplHealpixSkyPixellator
Fix the healpix.essentials-based HealpixSkyPixellator implementation
to use the HealpixBase.queryDiscInclusive method instead of queryDisc.
I think it did the wrong thing before.

Note this class is only compiled during tests, and it's not used
even then.  But one day it might be, if I move from the PixTools
HEALPix implementation to the HealpixBase one, and in any case
it could be useful for testing.