table: add FAILING tests for PixTools bug
Alexey Mints identified a bug in the Tilings.healpixRingIndex method,
which I traced back to the PixTools vect2pix_ring method.

The nature of the bug was that this PixTools method:

   vect2pix_ring(long nside, javax.vecmath.Vector3d vector)

returned the wrong pixel index for the phi<0 half of those pixels
straddling phi=0 in the equatorial region.
This applies to 2^k half-pixels out of the 12*4^k pixels at level k.

This commit adds some unit tests that are KNOWN TO FAIL in consequence
of this bug.  If the bug is fixed, the failures will go away.