table: remove 10240-char limit for output CSV fields
There was a limit in place for the maximum length of fields written
to output CSV files, namely 10240 characters.  This was there to
defend users from writing ridiculously long fields.  But it doesn't
provide much value, and it resulted in a GACS helpdesk ticket,
since Gaia DR2 has long array-valued fields which STIL serialises
(though does not not deserialise) like "(a,b,c,...)".

So remove this limit.  Add options that let you re-introduce it on
a custom basis if you instantiate your own CSV output handler.