ttools: fix issue with plot2 ifmt suffix wrangling
The parameter handling in AbstractPlot2Task looks for unsuffixed
versions of suffixed parameters so you can leave off suffixes
if they are the same for multiple layers.  That mostly works,
but it's extra complicated for parameters which have 'associated'
parameters like the TableInputParameter which manages its own
format and stream-flag parameters.  Modify the implementation
to get this right.  The resulting bug was reported by Peter Draper.