ttools: fix headless ErrorRenderer regression bug
When performing a headless plot to an external graphics format
(exporting a plot), and the plot uses ErrorRenderer (i.e. error
bars or ellipses etc), and the renderer's drawErrors() rather
than getPixels() method is used, then error bars were coming
out very small - only one or two pixels in extent.

This was a regression bug introduced between STILTS release
v3.1 and v3.1-1, at (bugfix) commit 5bbf961dd.

The problem is that in a headless context the call to try to
find out the size of the target graphics area (output image size)
doesn't seem to work - it gives a 1x1 pixel dimension.

I've worked around this by testing for tiny extents, and signalling
a larger image in that case.  Not great, but better than it was.

Here is an example plot that wasn't working:

   stilts plot2plane in=tgas_source.fits icmd='head 20' x=pmra y=pmdec \
          layer1=xyerror xerrhi1=pmra_error yerrhi1=pmdec_error \
          errorbar1=lines shading1=flat forcebitmap=false \
          out=x.png