ttools: add desigTo{Icrs,Ra,Dec} functions to func.Strings
New utility methods are added to decode IAU-style object designators
into ICRS coordinates, following the rules in the document at
http://cds.u-strasbg.fr/vizier/Dic/iau-spec.htx.

The functions are available in the expression language from the
uk.ac.starlink.ttools.func.Strings class, though I'm not sure
that's a totally appropriate parent class for them.

This was discussed on the topcat-user mailing list following a request
from Petr Skoda.  These functions need to be used with care,
but if there's no other positional information, it's better than nothing.
This has been spelled out in the user-visible function javadocs.