ttools: upgrade JEL to v2.1.2
At my request, Konstantin Metlov made an adjustment to the way
dynamic/static variable resolution gets done.
In previous versions, the existence of a column with a given
name prevented use of a static function with the same name.
In most cases this isn't necessary, because the function has
a different signature than the (argument-less) column.
From this version, table columns and supplied functions can have
the same name without causing problems.