ttools: update taplint OBS stage for REC-ObsCore-1.1
A change in ObsCore since the last PR is that the columns
obs_publisher_did and publisher_id have had their UCDs changed
from "meta.ref.url;meta.curation" to "meta.ref.uri;meta.curation".
This commit updates taplint accordingly.
Note however that this UCD is illegal UCD1+ syntax, since the
word "meta.curation" is P[rimary], so not allowed to go in
second place in a composed word.  Therefore I added a note when
checking this UCD that you're onto a loser whatever you do
(an Erratum to ObsCore or maybe UCDList would seem to be required).

Also added a comment in the code that there is an inconsistency
in REC-ObsCore-1.1 concerning pol_states mandatoryness, but there
is no associated behaviour change.