topcat: simplify synthetic column description annotation
Modify the way that the defining JEL expressions for synthetic columns
make their way into the ColumnInfo Description metadata slot.
It's desirable for this to happen when tables are exported,
so you can see where calculated values came from, but you don't really
want it to be visible within the Columns Window, since that information
is already available in the Expression metadata slot.

Previously the description was written directly into the Description
slot, and TOPCAT made arrangements to display something a bit different
(the "base description").  This worked but was complicated, and the
display of description could be confused with the display of expression
if you weren't careful (it fooled me at least once).

In this revision, the representation of the expression in the
column description has changed (it is now prefixed by the string
"Expression: "), and the implementation has also changed;
the description is not affected as long as the synthetic column is
within the topcat application (and still synthetic), but tables that
are serialized (so the column values are effectively frozen)
are first prepared using the TopcatUtils.getSaveTable() method
to copy the expression into the description field before export.
This is less baroque and uses fewer public members and lines of code.