frog: fix for multiple versions of nom.tam.fits
Make necessary changes to FROG classes to build with recent
versions of nom.tam.fits as well as starjava version.

This includes removing some functionality (logging HDU content
on save).  It wouldn't be too hard to bring it back, but since
frog is as far as I know moribund, it doesn't seem worth the effort.