fits: try to make sure HIERARCH convention is in use
Provide a method FitsConstants.configureHierarch() that sets
the current status of the nom.tam.fits library to using,
rather than ignoring, the ESO HIERARCH convention.

By default vanilla nom.tam.fits sets this false.
The nom.tam.fits bundled with starjava sets it true (since fd527d661ef),
because it's needed for WideFits.  There is no facility to manage
this setting on a per-call basis.

This commit fixes things so that HIERARCH is probably set true
when it needs to be, though it's hard to make this bulletproof
against the possibility that there are other classes in the same
JVM with different ideas.