topcat: fix subset stack scrollbar issue
When there are too many subsets to display in the plot window
Subsets tab, a scrollbar appears at the right of the subsets list.
But the subsets list stayed the same width, meaning that the right
hand side of the list got obscured by the scrollbar.

Implement a weird layout hack that I stole off StackOverflow and
don't understand.  This seems to fix it.  Now the subsets list
component gets wider when the scrollbar is present, so you can
still see its full width.

This issue was reported by Philippe Pognonec (who has lots of subsets
with 1-character names).