ttools: add new StorageType.LONG and LongCoord
The plot2 data system is now capable of dealing with long integer
valued data columns.

This was introduced to allow the HealpixPlotter to work with
HEALpix levels >13, for which the pixel indices can exceed 2^31.
In fact there are some other issues plaguing that upgrade,
so this change is not used right now, but it is in place should
long-valued input coordinates be required at some point in the
future.