ttools: improve VotLinter.createParser method
VotLinter can now provide an XMLReader that will both perform
validation on a SAX stream, and simultaneously feed SAX events
to a user-supplied ContentHandler.  This makes it easy to use
the VotLint code to do VOTable-sensitive validation at the same
time as whatever other XML processing you need (for instance
constructing a DOM).