ttools: improve efficiency of ContourPlotter smoothing
The smoothing in ContourPlotting is a convolution with a square
top-hat kernel.
This was implemented as a 2-d convolution which uses N*N operations
per pixel for a smoothing kernel linear dimension of N, and it
was slow for medium-large kernels (10 is already sluggish).
But this can be improved.  Since this kernel is separable, the
operation can be replaced with two 1-d convolutions, using 2*N
operations per pixel.  Much better.