ttools: ContourPlotter now caches smoothed grid in plan
The ContourPlotter now calculates the smoothed grid during the
plan calculation stage rather than the painting stage.
This means that the plan uses more memory, but only scaling with
the pixel grid size.  But it means that repaints with the same
smoothing and plot surface don't need to recalculate the smoothing,
and also prepares for the ReportMap to contain level information.