topcat: restore polymorphic ValueInfo value formatting
There is some code in PlasticStarTable that wraps a DefaultValueInfo
in another DefaultValueInfo so that changes to the value info
(e.g. in the Parameter window) in one table don't affect the
values in the parent one.  That makes sense (though is a little
paranoid), but it defeats polymorphism: it means that ValueInfo
implementations won't get to call their own subclassed implementations
of methods like formatValue.  Adjust the code so that they do.