votable: fix VOTable input handlers to pick up more of DOM
The TableStreamer and MultiTableStreamer SAX handlers that are
used by TableBuilder to stream VOTable documents create their
VOTables without reading the input SAX stream in its entirety.
For the case of TableStreamer this is so it can avoid possibly
wading through large amounts of data later in the stream when
it already has the table it's looking for, and for the case of
MultiTableStreamer it's so that it can provide earlier tables
to the client code one at a time as they are collected from the
stream rather than having to wait for all of them before handing
the whole lot over.

But the implementation of this means that VOStarTable objects
are created before the whole DOM is available.  That matters if
there are items in the XML document after the TABLE which
ought to form part of the VOStarTable's state.

That is typically the case for DataLink-style service descriptors,
which are in a separate RESOURCE element, typically after the
relevant TABLE element.  So previously, the VOTable input handlers
tended to miss picking up service descriptors for input VOTables.

This commit modifies the behaviour of the VOTable SAX handlers
so that more of the DOM has been assembled
before the constructed VOStarTables are handed off to client code.

For MultiTableStreamer, all the XML up to the start of the next TABLE
is now read, rather than all the XML up to the end of the curent TABLE.
In most cases (including the kind of thing intended by the DataLink
document), that will probably include the parts of interest.

For VOTableBuilder.makeStarTable, it now uses the new class
SingleTableReader which does read the whole DOM before providing
a StarTable.  You could argue it's better to bail out befor the
TABLE after the target one, but this is safer in case useful
metadata is a long way later in the SAX stream.

VOTableBuilder.streamStarTable can't do much about this, because
of the definition of the TableSink interface (metadata all comes
up front).  Just add some documentation about that.