votable: output handler writes ServiceDescriptor RESOURCEs
If a StarTable has ServiceDescriptor-typed items in its parameter
list, these are now written as RESOURCE elements to the
output VOTable document, as described in the DataLink document.

The RESOURCE element is written directly after the relevant TABLE
element, i.e. nested in the TABLE's own RESOURCE.
This is somewhat different to what is suggested (though not mandated)
by the DataLink 1.0 specification, which indicates writing
the TABLE and the service descriptor in two sibling resources.
But in the case of a multi-table VOTable document, that would
leave you with no good way to work out which service descriptor
corresponds to which table.

In an earlier version I tried writing sibling table and
service-descriptor resources, with each table-descriptor pair
nested within its own top-level resource.  That sounds reasonable,
but it leaves the TABLE elements nested two resources deep,
and it turned out that meant Aladin didn't find them.